Refactor CrewAI adapters by consolidating multiple adapter functions into a single file, eg `crewai.py`, for improved organization and maintainability. The previous individual adapter files for agents, orchestrators, and tools have been removed.
